+++ config/policies.js
@@ -77,6 +77,8 @@ module.exports.policies = {
     'uab/timesheet/edit': ['isBanned', 'isSocket', 'isAuthorizedAdminDirectorUab'],
     'uab/timesheet/get': ['isBanned', 'isSocket'],
     'uab/timesheet/view': 'isBanned',
+    'underwritings/*': ['isBanned', 'isSocket', 'isAuthorizedDirector'],
+    'underwritings/get': ['isBanned', 'isSocket'],
     'xp/*': ['isBanned', 'isSocket', 'isAuthorizedDirector'],
     'xp/get': ['isBanned', 'isSocket', 'isAuthorizedHost']
 };